requirements.txt
Flask==3.1.2
blinker==1.9.0
click==8.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.3
gunicorn==22.0.0
psycopg2-binary==2.9.11
scikit-learn==1.7.2
pandas==2.3.3
requestspython-dotenv
